json: clean up headers
