+++ src/trx/bson/enum.h
@@ -1,0 +1,9 @@
+#pragma once
+
+typedef enum {
+    BSON_PARSE_ERROR_NONE = 0,
+    BSON_PARSE_ERROR_INVALID_VALUE,
+    BSON_PARSE_ERROR_PREMATURE_END_OF_BUFFER,
+    BSON_PARSE_ERROR_UNEXPECTED_TRAILING_BYTES,
+    BSON_PARSE_ERROR_UNKNOWN,
+} BSON_PARSE_ERROR;
