+++ src/trx/bson/write.c
@@ -1,4 +1,5 @@
-#include <trx/bson.h>
+#include <trx/bson/write.h>
+
 #include <trx/debug.h>
 #include <trx/log.h>
 #include <trx/memory.h>
