+++ src/trx/bson/write.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <trx/bson/types.h>
+
+/* Write out a BSON binary string. Return 0 if an error occurred (malformed
+ * JSON input, or malloc failed). The out_size parameter is optional. */
+void *BSON_Write(const JSON_VALUE *value, size_t *out_size);
