+++ src/trx/json/base.c
@@ -1,4 +1,4 @@
-#include <trx/json.h>
+#include <trx/json/base.h>
 
 #include <trx/memory.h>
