+++ src/trx/json/write.c
@@ -1,4 +1,4 @@
-#include <trx/json.h>
+#include <trx/json/write.h>
 
 #include <trx/memory.h>
